+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Dependency updates (Kafka 4.1.1, Vert.x 5.0.6, Netty 4.2.9.Final, JMX Prometheus collector 1.5.0, OAuth 0.17.1).
 * Adding support for [JsonTemplateLayout](https://logging.apache.org/log4j/2.x/manual/json-template-layout.html).
+* Add support for TLS/SSL on the HTTP interface
+  Set `http.ssl.*` configurations to enable it.
 
 ## 0.33.1